requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.0.6
django-bootstrap5==24.2
django-crispy-forms==2.1
djangorestframework==3.15.1
et-xmlfile==1.1.0
gunicorn==22.0.0
idna==3.7
numpy==1.26.4
openpyxl==3.1.3
packaging==24.0
pandas==2.2.2
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
